fix: satisfy oxlint 1.76.0 on hookGeneratedPaths tests
CI (pinned oxlint 1.76.0) failed on five issues introduced by this PR:

- src/lib/worktrees.test.ts and src/lib/config.test.ts exceeded their
  max-lines caps. Split the new hookGeneratedPaths-only test cases into
  sibling files (worktrees.hookGeneratedPaths.test.ts,
  config.hookGeneratedPaths.test.ts), matching this codebase's existing
  convention of splitting oversized spec files by topic (e.g.
  worktrees.create.test.ts, config.promptFile.test.ts) rather than
  raising the cap.
- Two vitest/no-conditional-in-test violations: replaced a ternary
  mock-return with an extra if-branch (matching the file's existing
  if-based mock-dispatch style) and extracted the other's `&&` filter
  predicate into a named module-level helper, removing the conditional
  from test scope instead of suppressing it.
- unicorn/no-useless-undefined: `.catch(() => undefined)` -> `.catch(() => null)`.

Also removed dead code: `makeConfig`'s unused `defaults` override
parameter in worktrees.test.ts, which no test ever passed.

Corrected two claims in the PR description: the removal-failure
transparency guarantee is narrower than originally stated (a locked
worktree with only declared hook dirt still reaches a --force retry,
so the caller sees that error rather than the original), and the
"reproduces on a clean origin/main checkout" claim was based on a
stale, mis-pinned node_modules — a real npm ci install on main is
green.
