+++ src/lib/config.test.ts
@@ -7,12 +7,7 @@ import {
   setEnvironmentVariable,
   snapshotEnvironmentVariables,
 } from "../testHelpers/env.ts";
-import {
-  type Config,
-  type KnownRepository,
-  type ResolvedConfig,
-  resolveHookGeneratedPaths,
-} from "./config.ts";
+import type { Config, KnownRepository, ResolvedConfig } from "./config.ts";
 
 interface ConfigModule {
   loadConfig: () => Promise<Readonly<ResolvedConfig>>;
@@ -621,195 +616,6 @@ describe("loadConfig", () => {
     const actual = await loadConfig();
 
     expect(actual.defaults.hooks).toStrictEqual({});
-  });
-
-  it("leaves defaults.hookGeneratedPaths unset when omitted", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({
-        workspace: VALID_WORKSPACE(temporary),
-        defaults: { hooks: { prepareWorktree: "npm ci" } },
-      }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    const actual = await loadConfig();
-
-    expect(actual.defaults).toStrictEqual({ hooks: { prepareWorktree: "npm ci" } });
-  });
-
-  it("normalizes defaults.hookGeneratedPaths by trimming, stripping trailing slashes, and deduping", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({
-        workspace: VALID_WORKSPACE(temporary),
-        defaults: {
-          hookGeneratedPaths: [".agents/skills/", ".agents/skills", "  .claude/settings.json  "],
-        },
-      }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    const actual = await loadConfig();
-
-    expect(actual.defaults.hookGeneratedPaths).toStrictEqual([
-      ".agents/skills",
-      ".claude/settings.json",
-    ]);
-  });
-
-  it("carries hookGeneratedPaths onto the knownRepositories entry", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({
-        workspace: {
-          projectDir: temporary,
-          knownRepositories: [{ name: "repo-a", hookGeneratedPaths: [".claude/settings.json"] }],
-        },
-      }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    const config = await loadConfig();
-
-    expect(config.workspace.repositories[0]).toStrictEqual({
-      name: "repo-a",
-      hookGeneratedPaths: [".claude/settings.json"],
-    });
-  });
-
-  it("resolves the per-repo hookGeneratedPaths over the global defaults", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({
-        workspace: {
-          projectDir: temporary,
-          knownRepositories: [{ name: "repo-a", hookGeneratedPaths: [".agents/skills"] }, "repo-b"],
-        },
-        defaults: { hookGeneratedPaths: [".claude/settings.json"] },
-      }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-    const config = await loadConfig();
-
-    const actual = resolveHookGeneratedPaths({ config, repository: "repo-a" });
-
-    expect(actual).toStrictEqual([".agents/skills"]);
-  });
-
-  it("falls back to the global defaults when the repo entry omits hookGeneratedPaths", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({
-        workspace: { projectDir: temporary, knownRepositories: ["repo-b"] },
-        defaults: { hookGeneratedPaths: [".claude/settings.json"] },
-      }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-    const config = await loadConfig();
-
-    const actual = resolveHookGeneratedPaths({ config, repository: "repo-b" });
-
-    expect(actual).toStrictEqual([".claude/settings.json"]);
-  });
-
-  it("resolves to an empty list when neither layer declares hookGeneratedPaths", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      validConfigSource({ workspace: VALID_WORKSPACE(temporary) }),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-    const config = await loadConfig();
-
-    const actual = resolveHookGeneratedPaths({ config, repository: "repo-a" });
-
-    expect(actual).toStrictEqual([]);
-  });
-
-  const invalidHookGeneratedPaths: Array<[string, RegExp]> = [
-    ["/etc/passwd", /hookGeneratedPaths\[0\] must be a relative path inside the worktree/],
-    ["../outside", /hookGeneratedPaths\[0\] must not contain '\.' or '\.\.' segments/],
-    [".", /hookGeneratedPaths\[0\] must not contain '\.' or '\.\.' segments/],
-    [":(glob)**", /hookGeneratedPaths\[0\] must not start with ':'/],
-    ["   ", /hookGeneratedPaths\[0\] must be a non-empty string/],
-  ];
-  it.each(invalidHookGeneratedPaths)(
-    "rejects the hookGeneratedPaths entry %s",
-    async (entry, expected) => {
-      const configPath = writeConfigFile(
-        temporary,
-        validConfigSource({
-          workspace: {
-            projectDir: temporary,
-            knownRepositories: [{ name: "repo-a", hookGeneratedPaths: [entry] }],
-          },
-        }),
-      );
-      setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-      const { loadConfig } = await loadFreshConfig();
-
-      await expect(loadConfig()).rejects.toThrow(expected);
-      await expect(loadConfig()).rejects.toThrow(/workspace\.knownRepositories\[0\]\./);
-    },
-  );
-
-  it("rejects a non-string hookGeneratedPaths element", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      [
-        "export default {",
-        `  workspace: { projectDir: "/dev", knownRepositories: [{ name: "repo-a", hookGeneratedPaths: [1] }] },`,
-        `  agents: { definitions: { claude: {} } },`,
-        "};",
-      ].join("\n"),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    await expect(loadConfig()).rejects.toThrow(
-      /workspace\.knownRepositories\[0\]\.hookGeneratedPaths\[0\] must be a non-empty string/,
-    );
-  });
-
-  it("rejects a hookGeneratedPaths value that is not an array", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      [
-        "export default {",
-        `  workspace: { projectDir: "/dev", knownRepositories: [{ name: "repo-a", hookGeneratedPaths: "x" }] },`,
-        `  agents: { definitions: { claude: {} } },`,
-        "};",
-      ].join("\n"),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    await expect(loadConfig()).rejects.toThrow(
-      /workspace\.knownRepositories\[0\]\.hookGeneratedPaths must be an array/,
-    );
-  });
-
-  it("rejects a defaults.hookGeneratedPaths value that is not an array", async () => {
-    const configPath = writeConfigFile(
-      temporary,
-      [
-        "export default {",
-        `  workspace: ${JSON.stringify(VALID_WORKSPACE(temporary))},`,
-        `  defaults: { hookGeneratedPaths: "x" },`,
-        `  agents: { definitions: { claude: {} } },`,
-        "};",
-      ].join("\n"),
-    );
-    setEnvironmentVariable("GROUNDCREW_CONFIG", configPath);
-    const { loadConfig } = await loadFreshConfig();
-
-    await expect(loadConfig()).rejects.toThrow(/defaults\.hookGeneratedPaths must be an array/);
   });
 
   it("rejects non-object defaults", async () => {
